Envelope Trees and Bug Fixes
