+++ pkg/trisa/envelope/testdata/alice.vaspbot.net.pem
@@ -1,0 +1,120 @@
+-----BEGIN CERTIFICATE-----
+MIIF9DCCA9ygAwIBAgIUFTgH40htqvvzgt8KvahS8jb3NYAwDQYJKoZIhvcNAQEL
+BQAwcTELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlmb3JuaWExEzARBgNVBAcM
+Ck1lbmxvIFBhcmsxDjAMBgNVBAoMBVRSSVNBMRAwDgYDVQQLDAdUZXN0TmV0MRYw
+FAYDVQQDDA10cmlzYXRlc3QuZGV2MCAXDTI0MDMyODIxNDQ0M1oYDzIwNTQwMzIx
+MjE0NDQzWjB2MQswCQYDVQQGEwJVUzERMA8GA1UECAwITmV3IFlvcmsxETAPBgNV
+BAcMCE5ldyBZb3JrMRMwEQYDVQQKDApBbGljZSBWQVNQMRAwDgYDVQQLDAdUZXN0
+aW5nMRowGAYDVQQDDBFhbGljZS52YXNwYm90Lm5ldDCCAiIwDQYJKoZIhvcNAQEB
+BQADggIPADCCAgoCggIBAMSoGq0xtzwNmWjp2FrczP7+woqlPAuKmk+wraf4Z4B4
+rNu/kQ3O9ChqtcjVLz8qcC4zaFTfST5epUoLtYsS6e33K2dFgUFq/N6gWLEMXQJJ
+FqaAPDxHU2HPJtBuQrjHzQh0GexzXsMpl/rHnQzzvTP8IeA5HVxWZKh9ASdVuOIa
+/IBMSjQ+jMuJKUhdG83ui9c+Cx4HS5K2b28Ag7ODmc8EdOHflj3zfKvldLAW3m1i
+5Viq16eftJrPWCjgdgs55Du525zaOdAMZBGnHYj3yHK0t/LIoanyohbiFz7NFO+8
+O1uAAsVNSgXj14oY0oX7ohVYfWAe0OawfSERkKGGj9V94Qra0UvRsZx3/pjIwA2d
+4uu3XCehBfuI5p1mTINEV1/zYMt9qIop6YixgVm28TvGTurep0J+jHRBNGWoz/GU
+5t+fWQMvq/UZmveV1L+ehzOxsH98JccbDjtabNJwgfft2j0w0yGrPzQA96Hemllg
+QogkR2Nb8cPO6HAwl6VUv9+9yv/NIT7zElScMjjaHJFb38iUWhpCAmuM2tohhUvE
+PmMFX6r8FLpEkd7JR6Ve8Cyqq98Thzc4zqZxXrL2UjdhSYt4zqo1Y4M2MXoz2yqi
+/jHDTwpVUwmDH8TdggSyM6DJ4oG9/qpyjOaHVwOBd9mfgKA8KskIoy6fOsV1ge4H
+AgMBAAGjfTB7MDkGA1UdEQQyMDCCEWFsaWNlLnZhc3Bib3QubmV0ghMqLmFsaWNl
+LnZhc3Bib3QubmV0ggZidWZuZXQwHQYDVR0OBBYEFFP6/TEpDhIPQXCC9OIqMi2w
+mHBXMB8GA1UdIwQYMBaAFBYyksQmQJLfnS/Yn0CDC/NyEknNMA0GCSqGSIb3DQEB
+CwUAA4ICAQCW3DIB2VJ+k1vOjOvvnQk5y7V0C3C3kc0ZueA1ML5clR0y1xMxm1UO
+uKg/KIpmjVaQGS0CdK2tFft4wRtP3h7KOYN9EO71+ek/jDBb9ygUznylOiq8vjzn
+YUBp5Zb8QtEeCWT+BCTArZl0TkNbbKFlEKjDUuBJgcEz/XLtbtourURnJGlJp0BX
+R763C/vxuimHayuXU/ZYdxOypVQIXwyxPLiTlgQW0Ev22FUe5498T2SORNm0+8Ul
+WbNL49f5cjStfguv7yEWbqaXLiNNt1blIr9O8+eqGJDD4lakaALvCAn9BhE7XxJz
+6j/3ksKJkAM0I5SkOEFjka1hr+/iW6JYoc78+ku6h2gkfmYzWWLoRda9cj6a7roT
+TFbq+f4FLZOujYbAUYWdOtYo7uqnjQNytcO4z/kyI8lMgStkhlqUWEqhqWh/eV2D
+yu5gRjL45pT/ceSa5u1zLtVGPmkx5vjO/6Uc4Q/r5vb71favUP4KEZXlZKwyoEgw
+/JerQOfEQ8sdY49hKAAZbLdhrDpsJLWdloL7PT5/EF1CyckbclK2O6yQMsYAsxtc
+FtpiedcHpDi/XlFFGtjFRe04wGichBkYH0JAAmY0WPK0JfmlX6NJ9Mpmeba35f08
+5q/c0RkcbCpaMYq5kB8dUlmjCvyX+Yt7iC2FfZnN8bKVYdxUi+f7OQ==
+-----END CERTIFICATE-----
+-----BEGIN CERTIFICATE-----
+MIIF8DCCA9igAwIBAgIUbYA5ho2JT+MEH5crRmDjxREI7UowDQYJKoZIhvcNAQEL
+BQAwcTELMAkGA1UEBhMCVVMxEzARBgNVBAgMCkNhbGlmb3JuaWExEzARBgNVBAcM
+Ck1lbmxvIFBhcmsxDjAMBgNVBAoMBVRSSVNBMRAwDgYDVQQLDAdUZXN0TmV0MRYw
+FAYDVQQDDA10cmlzYXRlc3QuZGV2MCAXDTI0MDMyODIxNDQ0MloYDzIwNTQwMzIx
+MjE0NDQyWjBxMQswCQYDVQQGEwJVUzETMBEGA1UECAwKQ2FsaWZvcm5pYTETMBEG
+A1UEBwwKTWVubG8gUGFyazEOMAwGA1UECgwFVFJJU0ExEDAOBgNVBAsMB1Rlc3RO
+ZXQxFjAUBgNVBAMMDXRyaXNhdGVzdC5kZXYwggIiMA0GCSqGSIb3DQEBAQUAA4IC
+DwAwggIKAoICAQCtSwWp/hu5pYRmxC/L8le3bYmCks44HQncGy9y+jX9Ec+B10ir
+JKiBmunAqAir3N656hY1a7doBWajceIrov+jrUZYezQmhDawS1wEpKKgJAHUit+a
+4KC78wMSdSahAIIqaNsG93wezpPzKxE3e5mc9DqnPxHioMZIHIV05HTM4GSe22tG
+0rRw07/URP+sBXIO8aEbWOv6+oLS/cM6EGR44j4x8zn0J/gEOfSHUcj6eWojVfrN
+rdlufkydGoK1BrqlhC9NiSdLxXj/3yM0Lkbs36hvdFspQErVVcJ2Li2cs9upiGhT
+meU2tJuiHPR5WgBEaegzGUCL8V+t4WtC8KiLcAnWXdj1ugg4BVN0ab25CaKNy4Ze
+S5ND/HXlnrQVNavzKDJsknn+zwq9GZjGc5O3iKtgKn/tzhsBpAZKqiCUohGHcYMG
+/1aicc2QIdtYmdiMoJCLcUcZSGN0qSGPxSpUXexuBrxBCGnPo8wp1Gxp/Jtoocbo
+TAdzyc4A5lGYQ4gztymrs39xFlQs1BaSv603BL4A+mrG4/FNbRH47nzBw9KflOid
+uXMLwr3trKlbKV1hmK0Zy+T84+iEnIEQ9i73Ak10n1ouVR2FgaYr7XSFsBZ8Ymm5
+vBYnqAh4Danv8fjdiKtXAgFE54LE8L2AVQVrnL9jslHLuCmbQ8GIN8tLCwIDAQAB
+o34wfDAdBgNVHQ4EFgQUFjKSxCZAkt+dL9ifQIML83ISSc0wHwYDVR0jBBgwFoAU
+FjKSxCZAkt+dL9ifQIML83ISSc0wDwYDVR0TAQH/BAUwAwEB/zApBgNVHREEIjAg
+gg10cmlzYXRlc3QuZGV2gg8qLnRyaXNhdGVzdC5kZXYwDQYJKoZIhvcNAQELBQAD
+ggIBACSe//FHwsejGg/U2E7Adr4Ody/k3Pue6sAa09ioPxPYs4bGhTnL34xJDJKh
+AIRMV8XxLJeZp5SsAP/8lp2rxUDphJP2ePuo/6aK8ZEZDK0kEi8dbzs95UTxJgKH
+NofjU15yc2309vMqjCeNwbtnhNHXckYkeefwg5GUuc1CiiNOGEN70LW9kZxu5kf5
+0DMK7DwxC2bg8UdJj0oNxR2Ijs08SowT5j+xiwCzlF4sdvAyYyDsra9+3LvDa6Uy
++iuR+LHD2JjzcW9hgbhKLHyaadweyh/6Cp7nsC+u6E6T0rYhF/swZR17h6FYhDU4
+Uh89eMa4zioH1t/Zd3DMQl/dm13/w+n0sfE8II/JUWLwF7Pn92CyiDz09hHT4R/W
+v6NuMc5G0ePdAO+SMroMyqr9yuZePTBA5mCeWOaywFGsMEuTv/6LAntGg5Tn+5tk
+BKsbK7r9LPbDtqnW0RmPmyLZxMgpLhFeDDleK2XpJzunidHezxYpCdLqAQF3n25I
+Ny8lRkHd4uSr0VITrkdYT0c9mDpQ/mOjd2roI63C2PXbGiH913aSG2gt4BvELtjF
+bSli5/naN4erKsTyQ+/sp+uOrHt310T5ipYl8m4ZB4cbEftODr/+JQbcVit5/uF2
+3YQYlZKdvRlmQ87TFL3Zu39y24ljiEyQ+Xl76gJ8U1LJrBQi
+-----END CERTIFICATE-----
+-----BEGIN PRIVATE KEY-----
+MIIJQgIBADANBgkqhkiG9w0BAQEFAASCCSwwggkoAgEAAoICAQDEqBqtMbc8DZlo
+6dha3Mz+/sKKpTwLippPsK2n+GeAeKzbv5ENzvQoarXI1S8/KnAuM2hU30k+XqVK
+C7WLEunt9ytnRYFBavzeoFixDF0CSRamgDw8R1NhzybQbkK4x80IdBnsc17DKZf6
+x50M870z/CHgOR1cVmSofQEnVbjiGvyATEo0PozLiSlIXRvN7ovXPgseB0uStm9v
+AIOzg5nPBHTh35Y983yr5XSwFt5tYuVYqtenn7Saz1go4HYLOeQ7uduc2jnQDGQR
+px2I98hytLfyyKGp8qIW4hc+zRTvvDtbgALFTUoF49eKGNKF+6IVWH1gHtDmsH0h
+EZChho/VfeEK2tFL0bGcd/6YyMANneLrt1wnoQX7iOadZkyDRFdf82DLfaiKKemI
+sYFZtvE7xk7q3qdCfox0QTRlqM/xlObfn1kDL6v1GZr3ldS/noczsbB/fCXHGw47
+WmzScIH37do9MNMhqz80APeh3ppZYEKIJEdjW/HDzuhwMJelVL/fvcr/zSE+8xJU
+nDI42hyRW9/IlFoaQgJrjNraIYVLxD5jBV+q/BS6RJHeyUelXvAsqqvfE4c3OM6m
+cV6y9lI3YUmLeM6qNWODNjF6M9sqov4xw08KVVMJgx/E3YIEsjOgyeKBvf6qcozm
+h1cDgXfZn4CgPCrJCKMunzrFdYHuBwIDAQABAoICAE+hiGfQRVO2aAPhCQBF/2ZT
+75+np+g+gBy7vJ3TCsotU0WKTSjLv/crup9vn7mSrCkxHNKdNbHhLkoM8r61cm8v
+Em63aM7+DRXy1OcgS/s0cE0MiReZhCyLbrqgozjiguYk6ShjUSBy070zHieT/O2z
+b141kmuE+i51q8VrQzmfVtZ2zedY2rdCO6q3NR6OtSZ705Gdv57Ra88FZM8If3wh
+4FhkKpi8YyPR/o2dPQULMjZUu41/MGktg5PNzE7gasg6irB5d0aE9fJl+qIvP4Yf
+IotXbYr9GmBsvZ/q+ErZLCnxbG3LTamT6H/dbEXfrnK5Sc9luw5mscx1qdyWOeAx
+3N56nnlFjDelfV0SY8+ysTYKWvUt5UjhLYj9qryL50RAjOQArUqgjy7kR/M3vd8e
+Vt7lZzZpW0rysiTkOkwBh098K2MZuGpDu0pDRnEJ702XHoF4/PkP07+JEY0dwYqB
+GNt54uq2AWHizzCrvefrEkK3CcdP+DmsHz5TUc8pK45LgQNXmJhEelEiPXt0EQq0
+dWmoXA1YGs4hQwgPnmV7KGUDr7XJlYndFNa18kbDa3mdSYZjiLNOM81Ngkb+kELw
+OjVdF2eS45/URBL0yK8xMDdmMtkUdgy1KQDc4sqKey5M5AFoOo2K3C10WD9KZ3uf
+2yRlRtQKcV4pS69ljGidAoIBAQDlvJGD9hOv5pY0rPtszrdGUn1gBoZEd3knciMI
+KgwwDS6ze6AA0osA3n10ZCgAcUwcxn6xnlg2AgwCwh1QYLadW6R0Y0x/dSKGCPjO
+QPMpraZ3IZYfkaoYkNO+j3gJqFabrEVHYX7PnCpAO2fIhOnn4aQVTfkMmJizi6CA
+rEZ2tYZoFr3Hi0WmWzT+AigNa2jTf7fHx675wJaftNWpFsVxh0Dd/WtbK8n8l/IL
+QIQTtTVjf6RTpb8p1DRvr8PRpAI19Z8M3ZHK7Qt35IDkebphBV48iKZWlasbomxw
+Ousty5IZvOgjISbFbgpE3TQYxl5f39ph3xpzTMeLoTzd8t1bAoIBAQDbI2yMPyBr
+jo6Rt7hJGVYp3lcu/LEmW11RfZBHPt4tYzgk2WZ9vqTVqxaghS+X9RUiBoM9FMNc
+7BZGm03xGNFJ7IIyZNOrv2B/Zz2x5jA1qROeN7f/2FlVJ4zY04kqhxIqxiZOMz19
+rAofVwhuLoUrttcl9mW0wRxJQfXKiI8xXKe5N3+OpuZ7p82tlm9QGscYkFFCv0UB
+4gIwKuaBmfLPRmyaBzP3t7i0iKP0epqqxidEmIMlzCuphcwNCqVCstpCNCCFTiVn
+AKP5g2JTYfWsU/j3Ws2Eg7LyNlwZODD9abHRf1rompPzfIlmHlfsUzaeU83U0uMt
+9FVAriz6yXXFAoIBAEzqo3WhD9pWw3eDavJ1C1uaBqv9wzptHb0dM6lqGoaEA1zI
+STu9QhjaZPpxBguP40HHG2fwcewzJz5NK29b0ghBRIBLNrN9zj8+Bb9Yc5FCuHcu
+YYrURDTRWHN2qWPiXozkUpWhiMmNqX+z4/14sq+WFk+juXyEIqwKVYR/KWBZSlTT
+OFr0wC8AXm+k3TARBBm7qxZSPr0Tw9pYuyhPnW9zLz6juCvgL1JItRsbUJ0gkG1t
+sODon2YrzBqQqGkqFitmvweZr1RXpP1RHe5g6wvMtk5iGf7nQVCRQukYcOD2RUYk
+vzvzv31eaEXCIc0hrTtAQWd/QOXVkQGozC4rP/0CggEAfVRzPnrUJe+wZnK2uUf+
+WY/KORtYjeFvK64umDDCjR7T+29DCOPCDln9ZO0HXVdUMNOct5Bqc3iq/NLR7vQM
+rsTPadb0oKOhovv+8wH9zJLYn4Kqf27TGLq3+UJyjpoVr9UID22K25dLasUyEyIU
+E/5Mam/Tl19iuBs7YgxcKRUe7/VnxMR7yXkdTwxcuWm3OLtBXnVaGEuUiMvgbXsI
+vbc/YZCBDkpLHyWO78I3NziBOOApEbMFvbzCCStvfQghf/+kIdmh1pktLwUPdTTv
+dxHHsGCEbieMbq7cWagjjKuogXLfIW1W9MjjJD5ydrzY3hB8Bh5ew+eb4pQ7MWuL
+KQKCAQEA4DyRKz4TyER5W+ytit4LSND6eZzdFgONQ+hMLiGrTN2FkSZsCycIDOyO
+yyb8339JVC2jh9U38ZGgIBLv6prcm2H27xH779zN/Z+1PSt6BKrlYFVwsjjma6iI
+ssDUokCrZ+5FBHI/WMCWg9O05dXCjMdMCNK0tp/k+/8sTqSiwydtRbJQpmcujyLx
+VMXqr72Z4rVm4A/HR8Z5yjK/C5aEXAxDOKriMDlYoHkR8+MDmnjE3hTk/Um1J5ly
++Mwaf8ZbejFxJqZeEtkDj0JpOMOrkaEf7p7xnfxF+VmHxDfz/1APICF+xA3g8c9h
+fUEUQwvHKHfotwZpvkPUrBJ3btmKkQ==
+-----END PRIVATE KEY-----
